completed user authorization

+++ app/controllers/api/users_controller.rb
@@ -3,12 +3,17 @@ module Api
   class UsersController < Api::BaseController
     before_action :authenticated?
 
+    after_action :verify_authorized
+
     def index
+      authorize current_user
       render json: User.all
     end
 
     def create
       user = User.new(user_params)
+      authorize user
+
       if user.save
         render json: user, status: :created
       else
@@ -19,15 +24,13 @@ module Api
 
     def destroy
       user = User.find(params[:id])
+      authorize user
 
       if user.destroy
         destroy_successful
       else
         destroy_error
       end
-
-    rescue ActiveRecord::RecordNotFound
-      object_not_found
     end
 
     private
